% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDetMod.R
\name{BayesDet}
\alias{BayesDet}
\title{Deterministic Bayesian inference on lineage growth curves.}
\usage{
BayesDet(area, times, model_specification, it, th, ch)
}
\arguments{
\item{area}{- Area Estimates on which to do inference in a matrix}

\item{times}{- Times associated with Area Observations in a matrix}

\item{model_specification}{- Model type. 1=logistic; 2=exponential;
3=log-linear; 4=exponential with B; 5=log-linear with B}

\item{it}{- Number of Iterations}

\item{th}{- Thinning}

\item{ch}{- Number of Chains}
}
\value{
Returns the mean parameter estimates for all input time courses and the MCMC chain.
}
\description{
Deterministic Bayesian inference on lineage growth curves.
}

