% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHybridMod.R
\name{BayesHybrid}
\alias{BayesHybrid}
\title{Bayesian Inference for Hybrid Model.}
\usage{
BayesHybrid(params, pmin, pmax, iters, tune, thin, noiseSD, switchN, dat)
}
\arguments{
\item{params}{- Initial guess for Growth Rate r and Carrying Capacity k in a vector}

\item{pmin}{- Lower bound of prior for r and K}

\item{pmax}{- Upper bound for prior for r and K}

\item{tune}{- Tuning Parameter}

\item{thin}{- Thinning}

\item{noiseSD}{- Precision Measures (Standard Deviation)}

\item{switchN}{- Number of cell after which to switch from the stochastic to the deterministic model}

\item{dat}{- Growth curve on which to no inference. Vector of cell counts with time points specified as rownames}
}
\value{
MCMC chain for r and K.
}
\description{
Bayesian Inference for Hybrid Model.
}

