% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqDetExpMod.R
\name{DensityOverlay}
\alias{DensityOverlay}
\title{Overlaying Growth Rate Density Distributions.}
\usage{
DensityOverlay(rates, data, maxr, datsetname)
}
\arguments{
\item{rates}{- Vector with all growth rates for that data sets}

\item{data}{- Data for that dataset; dataframe where the first column specifies the genotype}

\item{maxr}{- Maximum growth rate value displayed on the y-axis}

\item{datsetname}{- Data set for which the growth rates are specified}
}
\value{
Plot the growth rate density distributions for a dataset.
}
\description{
Overlaying Growth Rate Density Distributions.
}

