% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHybridMod.R
\name{detLog}
\alias{detLog}
\title{Logistic Growth Model.}
\usage{
detLog(K, r, N0, t)
}
\arguments{
\item{K}{- Carrying Capacity}

\item{r}{- Growth Rate}

\item{N0}{- Population Start Size}

\item{t}{- Time points for which to do inference; single value or vector}
}
\value{
A vector of cell count(s) for the specified input time(s) is returned.
}
\description{
Logistic Growth Model.
}

