% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detstocgrowth.R
\docType{package}
\name{detstocgrowth}
\alias{detstocgrowth}
\alias{detstocgrowth-package}
\title{detstocgrowth: A package for analysing microbial cell growth.}
\description{
The detstocgrowth function provide four categories of important functions:
loglin, popsim, det, stoch
}
\section{loglin functions}{

The login function assume exponential growth an apply a log-linear regression to the data.
}

\section{popsim functions}{

The popsim functions take single lineage growth rates as an input and return corresponding population simulations.
}

