% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqDetExpMod.R
\name{EXP_growthrate}
\alias{EXP_growthrate}
\title{Growth rates estimates inferred using the exponential model on the original scale.}
\usage{
EXP_growthrate(ai, ti)
}
\arguments{
\item{ai}{- Area vector where each row consists of one growth curve.}

\item{ti}{- Time vector where each row corresponds to one growth curve.}
}
\value{
Estimated growth rate (\code{rate}) and
 starting population (\code{intercept}).
}
\description{
Growth rates estimates inferred using the exponential model on the original scale.
}

