% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{lagduration}
\alias{lagduration}
\title{Calculating the duration of the lag phase for various population start sizes.}
\usage{
lagduration(strain = 0, k, t, N, it)
}
\arguments{
\item{strain}{- Strain data as returned by the \code{subset_strain} function or else \code{strain=0}}

\item{k}{- Growth Rates}

\item{t}{- Simulated Time Series}

\item{N}{- Vector of population start sizes for which to estimate lag duration}

\item{it}{- Number of Iterations}
}
\value{
Returns a list with the mean lag lengths \code{m}, the upper confidence intervals \code{cu},
the lower confidence intervals \code{cl}.
}
\description{
Calculating the duration of the lag phase for various population start sizes.
}

