% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqDetExpMod.R
\name{LM_growthrate}
\alias{LM_growthrate}
\title{Log-linear regression of the exponential model.}
\usage{
LM_growthrate(ai, ti)
}
\arguments{
\item{ai}{- Area vector}

\item{ti}{- Time vector}
}
\value{
Estimated growth rate (\code{rate}) and
 starting population (\code{intercept}) as well as the model fit (\code{fit})
 which can be accessed using the \code{$} operator.
}
\description{
Log-linear regression of the exponential model.
}

