% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSimAnalyses.R
\name{meanvarsim}
\alias{meanvarsim}
\title{Population Growth Rates with Increasing Start Sizes}
\usage{
meanvarsim(strain, strain_rates, N, iterations, t)
}
\arguments{
\item{strain}{- Strain Data from which to simulate}

\item{strain_rates}{- Strain rates from which to simulate}

\item{N}{- Largest population size}

\item{iterations}{- Number of iterations}

\item{t}{- Time sequence over which to simulate}
}
\value{
Returns two lists with the lag and exponential growth rate .
}
\description{
Population Growth Rates with Increasing Start Sizes
}

