% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqDetExpMod.R
\name{piecewise_LM}
\alias{piecewise_LM}
\title{Piece-wise log linear regression of the exponential model.}
\usage{
piecewise_LM(ai, t)
}
\arguments{
\item{ai}{- Area matrix where each row consists of one growth curve.}

\item{ti}{- Vector of chosen time intervals for simulating.}
}
\value{
Estimated slope before (\code{slope1}) and after (\code{slope2}) the
 breakpoint (\code{bp}) along with the estimated starting population
 (\code{intercept}) and fit (\code{fit}) which can be accessed using the \code{$} operator.
}
\description{
Piece-wise log linear regression of the exponential model.
}

