% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{pin_pop_plot}
\alias{pin_pop_plot}
\title{Pin Population Estimates}
\usage{
pin_pop_plot(strainname, sa, st)
}
\arguments{
\item{strainname}{- Plot title}

\item{sa}{- Pin Area Estimates}

\item{st}{- Pin Time Points}
}
\value{
Plots the pin population data with a line of best fit assuming a lag, esponential and stationary phase.
The two break points and the three consecuive slope estimates of the fit are returned.
}
\description{
Pin Population Estimates
}

