% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqDetPRMod.R, R/PopSims.R
\name{plot_growth}
\alias{plot_growth}
\title{Plotting growth curves on the log scale.'}
\usage{
plot_growth(a, t, s, Nsample = 100, title = TRUE, hist = FALSE)

plot_growth(a, t, s, Nsample = 100, title = TRUE, hist = FALSE)
}
\arguments{
\item{a}{- Area}

\item{t}{- Time}

\item{s}{- Identifier}

\item{Nsample}{- No. of growth curves sampled}

\item{title}{- Adds \code{"Identifier; Nsample"} as title}

\item{hist}{- Adds a histogram of the estimated growth rates for all areas supplied using \code{LM_growthrate}}

\item{a}{- Area}

\item{t}{- Time}

\item{s}{- Identifier}

\item{Nsample}{- No. of growth curves sampled}

\item{title}{- Adds \code{"Identifier; Nsample"} as title}

\item{hist}{- Adds a histogram of the estimated growth rates for all areas supplied using \code{LM_growthrate}}
}
\value{
A time versus area plot of sampled growth curves and (if \code{hist=TRUE}) a histogram of all growth rates.

A time versus area plot of sampled growth curves and (if \code{hist=TRUE}) a histogram of all growth rates.
}
\description{
Plotting growth curves on the log scale.'

Plotting growth curves on the log scale.'
}

