% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{plot_pop_sim_dat}
\alias{plot_pop_sim_dat}
\title{Simulating multiple population growth curves.}
\usage{
plot_pop_sim_dat(strain = 0, k, t, it, yl, col, colrange)
}
\arguments{
\item{strain}{- Strain data as returned by the \code{subset_strain} function or else \code{strain=0}}

\item{k}{- Growth Rates}

\item{t}{- Simulated Time Series}

\item{it}{- Number of Iterations}

\item{yl}{- Limits in y=axis}

\item{col}{- Vector of colours}

\item{colrange}{- Vector of k values which are to match the colour of vectors provide}
}
\value{
Plots the population data as simulated by \code{pop_sim_dat}.
}
\description{
Simulating multiple population growth curves.
}

