% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqDetPRMod.R
\name{PlotPinObsSim}
\alias{PlotPinObsSim}
\title{Comparin Pin Population Estimates with Pin Population Simulations.}
\usage{
PlotPinObsSim(area, times, sortedfolders, obs, conv, iterations, leg = FALSE)
}
\arguments{
\item{area}{- Strain Pin Area Estimates}

\item{times}{- Strain Pin Time Points}

\item{sortedfolders}{- Pin Folder Names}

\item{obs}{- Number of Observations to Plot and Simulate}

\item{conv}{- Cell to Area Conversion Factor}

\item{iterations}{- Number of Population Growth Curves to simulate}

\item{leg}{- if TRUE legend is added to the plot}
}
\value{
Plots the pin population data with a line of best fit assuming a lag and the simulated population data overlayed.
}
\description{
Comparin Pin Population Estimates with Pin Population Simulations.
}

