% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHybridMod.R
\name{PlotPosteriorPredictive}
\alias{PlotPosteriorPredictive}
\title{Plotting the Posterior Predictives overlayed on the Growth Curve.}
\usage{
PlotPosteriorPredictive(data, parmat, switchN, guessK)
}
\arguments{
\item{data}{- The growth curve data as supplied to \code{BayesHybrid}}

\item{parmat}{- MCMC chain returned from \code{BayesHybrid}}

\item{switchN}{- Number of cell after which to switch from the stochastic to the deterministic model}

\item{guessK}{- inital guess for K}
}
\value{
Plots the original growth curves, the posterior predictives and the posterior predictives overlayed on the data.
}
\description{
Plotting the Posterior Predictives overlayed on the Growth Curve.
}

