% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHybridMod.R
\name{PlotPosteriorProb}
\alias{PlotPosteriorProb}
\title{Plotting the Posterior Probabilities for an MCMC chain.}
\usage{
PlotPosteriorProb(pmin, pmax, parmat)
}
\arguments{
\item{pmin}{- Lower bound of prior for r and K}

\item{pmax}{- Upper bound for prior for r and K}

\item{parmat}{- MCMC chain returned from \code{BayesHybrid}}
}
\value{
Plots the posterior probability distributions.
}
\description{
Plotting the Posterior Probabilities for an MCMC chain.
}

