% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSimAnalyses.R
\name{pop_comoposition}
\alias{pop_comoposition}
\title{FishPlots for Population Simulations}
\usage{
pop_comoposition(r, t, gr)
}
\arguments{
\item{r}{- Sampled Growth Rates}

\item{t}{- Time sequence over which to simulate}

\item{gr}{- Range of Growth Rates}
}
\value{
A FishPlot for the fastest growing strains
}
\description{
FishPlots for Population Simulations
}

