% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSimAnalyses.R
\name{pop_composition_dat}
\alias{pop_composition_dat}
\title{Calculates Population Compositions as fed into the Fishplots using \code{pop_comoposition}}
\usage{
pop_composition_dat(t, iter, N0, strain_rates)
}
\arguments{
\item{t}{- Time sequence over which to simulate}

\item{iter}{- Number of Iterations}

\item{N0}{- Population Start Size}

\item{strain_rates}{- List of Growth Rates}
}
\value{
The number of lineages which make up more than five percent and the percentage with which
the fastest lineage dominates the population are returned in a data frame.
}
\description{
Calculates Population Compositions as fed into the Fishplots using \code{pop_comoposition}
}

