% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqDetPRMod.R, R/PopSims.R
\name{pop_plot_growth}
\alias{pop_plot_growth}
\title{Plotting simulated population growth curves on the log scale.'}
\usage{
pop_plot_growth(a, t, s, title = TRUE, hist = FALSE)

pop_plot_growth(a, t, s, title = TRUE, hist = FALSE)
}
\arguments{
\item{a}{- Area}

\item{t}{- Time}

\item{s}{- Identifier}

\item{title}{- Adds \code{"Identifier; Nsample"} as title}

\item{hist}{- Adds a histogram of the estimated growth rates for all areas supplied using \code{LM_growthrate}}

\item{a}{- Area}

\item{t}{- Time}

\item{s}{- Identifier}

\item{title}{- Adds \code{"Identifier; Nsample"} as title}

\item{hist}{- Adds a histogram of the estimated growth rates for all areas supplied using \code{LM_growthrate}}
}
\value{
A time versus area plot of simulated growth curves and (if \code{hist=TRUE}) a histogram of all growth rates.

A time versus area plot of simulated growth curves and (if \code{hist=TRUE}) a histogram of all growth rates.
}
\description{
Plotting simulated population growth curves on the log scale.'

Plotting simulated population growth curves on the log scale.'
}

