% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{pop_sim_dat}
\alias{pop_sim_dat}
\title{Simulating multiple population growth curves.}
\usage{
pop_sim_dat(strain = 0, k, Nsample, t, it)
}
\arguments{
\item{strain}{- Strain data as returned by the \code{subset_strain} function or else \code{strain=0}}

\item{k}{- Growth Rates}

\item{t}{- Simulated Time Series}

\item{it}{- Number of Iterations}

\item{Nsamples}{- Simulated population Start Size}
}
\value{
A list of the simulated population data \code{PopData}, all simulated strain data
\code{total_simdata}, all sampled lineage growth rate \code{SimRates}, and their corresponding
growth curve indices \code{total_indices}.
}
\description{
Simulating multiple population growth curves.
}

