% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSimAnalyses.R
\name{pop_sim_plot}
\alias{pop_sim_plot}
\title{Simulated Population Growth Curves}
\usage{
pop_sim_plot(strain, strain_rates, gr, yl, it, t)
}
\arguments{
\item{strain}{- Strain Data from which to simulate}

\item{strain_rates}{- Strain rates from which to simulate}

\item{gr}{- Growth Rate range applied to colour-scale}

\item{yl}{- Maxmimum limits on the y axis}

\item{it}{- Number of iterations}

\item{t}{- Time sequence over which to simulate}
}
\value{
The number of lineages which make up more than five percent and the percentage with which
the fastest lineage dominates the population are returned in a data frame.
}
\description{
Simulated Population Growth Curves
}

