% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHybridMod.R
\name{simCellsHybrid}
\alias{simCellsHybrid}
\title{Function to simulate growth curves according to the Hybrid Model}
\usage{
simCellsHybrid(K, r, N0, NSwitch, detpts = 100)
}
\arguments{
\item{K}{- Carrying Capacity}

\item{r}{- Growth Rate}

\item{N0}{- Population Start Size}

\item{detpts}{- Time interval used for the deterministic model}

\item{Nswitch}{- Switch from stochastic to deterministic model}
}
\description{
Function to simulate growth curves according to the Hybrid Model
}

