% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHybridMod.R
\name{simCellsStoch}
\alias{simCellsStoch}
\title{Function to simulate growth curves stochastically}
\usage{
simCellsStoch(K, r, N0)
}
\arguments{
\item{K}{- Carrying Capacity}

\item{r}{- Growth Rate}

\item{N0}{- Population Start Size}
}
\description{
Function to simulate growth curves stochastically
}

