% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqDetExpMod.R
\name{simExponential}
\alias{simExponential}
\title{Applying the exponential model.}
\usage{
simExponential(r, t, N0 = 1)
}
\arguments{
\item{r}{- Growth Rate}

\item{t}{- Time}

\item{N0}{- Stating Population (default N0=1)}
}
\value{
A list of population estimates for each time point.
}
\description{
Applying the exponential model.
}

