% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSimAnalyses.R
\name{subset_identifier}
\alias{subset_identifier}
\title{Subsetting the formatted data according to strain.}
\usage{
subset_identifier(d, a, t, identifier)
}
\arguments{
\item{d}{- Data}

\item{a}{- Area}

\item{t}{- Times}

\item{identifier}{- Identifier Name}
}
\value{
A list of all identifier information: its \code{area, times, data, name} and \code{indices} which can
accessed using the \code{$} operator.
}
\description{
Subsetting the formatted data according to strain.
}

