% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{subset_pin}
\alias{subset_pin}
\title{Subsetting the formatted data according to pin.}
\usage{
subset_pin(d, a, t, pin)
}
\arguments{
\item{d}{- Data}

\item{a}{- Area}

\item{t}{- Times}

\item{pin}{- Pin ID}
}
\value{
A list of all pin information: its \code{area, times, data, name} and \code{indices} which can
accessed using the \code{$} operator.
}
\description{
Subsetting the formatted data according to pin.
}

