% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{subset_strain}
\alias{subset_strain}
\title{Subsetting the formatted data according to strain.}
\usage{
subset_strain(d, a, t, strain)
}
\arguments{
\item{d}{- Data}

\item{a}{- Area}

\item{t}{- Times}

\item{strain}{- Strain Name}
}
\value{
A list of all strain information: its \code{area, times, data, name} and \code{indices} which can
accessed using the \code{$} operator.
}
\description{
Subsetting the formatted data according to strain.
}

