% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSimAnalyses.R
\name{yellowredmap}
\alias{yellowredmap}
\title{Yellow to Red Map for Growth Rate}
\usage{
yellowredmap(gr)
}
\arguments{
\item{gr}{- Range of Growth Rates}
}
\value{
A yellow to red colour map which provides a measure for growth rate
}
\description{
Yellow to Red Map for Growth Rate
}

